requirements.txt
# Data Processing
numpy==2.0.1
pandas==2.2.2


# Web Scraping
playwright==1.40.0
requests==2.31.0
fake-useragent==1.4.0
lxml==4.9.3

# API Framework - FastAPI
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - PostgreSQL with SQLAlchemy
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.13.0




# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# HTTP Client
httpx==0.25.2

# Rate Limiting
ratelimit==2.2.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0

# Data Export
openpyxl==3.1.2
